Update Homie (#1)
I needed to be able to use a username that had 40 some characters in it. So I forked the repo and increased the max size to 64 characters. I needed to point PlatformIO at this new location.

I've added a Dockerfile to turn this project into a container.
In order to receive the MQTT commands, I had to write an entrypoint that connected to the MQTT network as well as connecting with vault to generate the new MQTT credentials.

* Get working.

* Update homie to 2.1 (pre-beta) and get TLS communication working.

* PlatformIO container for building ESP firmware.

* Use SciFiFarms' Homie repo.