HAL: Homie implementation that allows custom settings to control nodes and sensors (#2)
H.A.L. (Homie Assisting Life) is a Homie implementation that allows the JSON config file to dynamically setup and control HomieNodes.

It does this by having the code for setting up, reading, and controlling a given sensor all ready present. Just put in the needed settings for a sensor, and it'll run the code for that sensor.

The steps to add a sensor type are:

    Put needed #includes with the other #includes.
    Define pointers for HomieNodes, any sensor/switch objects, and a HomieSetting object.
    Configure HomieSetting object in setup()
    Add code to create nodes in setup(). You'll need to include the advertise commands here as well.
    Add code to setup sensors/switches, as well as setting the property in setupHandler().
    Finally, the code to actually read sensors/flip switches should go in the loopHandler().

I'm using global pointers for managing the HomieNodes. My hope is that this will allow me to have many types of sensors supported, and all that gets added is a bit of code, the sensor library, and a pointer. That way I'm not putting HomieNodes into memory that aren't needed. If the node is actually needed, a HomieNode will get dynamically created. At least, that's the idea.

Currently, I'm creating at least one HomieNode for every sensor/switch HAL supports. This seems fairly inefficient, and I figure there has to be a fairly low limit to how many HomieNodes an ESP8266 can support. I suspect it's more than the number of sensors an ESP8266 can physically support.
I tried to dynamically create the HomieNodes based off of the config, but the HomieNodes have to be setup before Homie.setup() has ran, and the config isn't available until after Homie.setup() has ran. This shouldn't be a problem in Homie 2.1 as it sounds like I'll be able to call Homie.setupConfig() (or something like that) before calling Homie.setup().

I added some libraries to the Dockerfile so that the building of firmware doesn't take as long. I tried to prebuild the firmware too, but I ran into issues with PlatformIO using the wrong version of Homie.

* Rewrite HAL to better facilitate adding more sensors.

* Create the /workespace/data/homie folder as part of the build.

* Remove the Homie lib install from the Dockerfile. Was causing more problems than it was worth.

* Have Homie config JSON dynamically set up the HomieNode objects and communications.

* Return to using the master branch.